+++ docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This a informal changelog so i can keep track of what im doing.
 
+## 2026-04-21
+
+- Added a button in settings to mass delete orphaned cbz files. (because tracking downloaded files properly is too annoying)
+
 ## 2026-04-19 - 2026-04-20
 
 - Added Suggestions Page with working suggestions!